refactor(040): migra packages/storage para TypeScript
